+++ backend/dataall/db/api/environment.py
@@ -29,7 +29,7 @@ from dataall.utils.naming_convention import (
     NamingConventionService,
     NamingConventionPattern,
 )
-from dataall.core.group.services.group_resource_manager import EnvironmentResourceManager
+from dataall.core.group.services.environment_resource_manager import EnvironmentResourceManager
 
 log = logging.getLogger(__name__)
