+++ backend/dataall/modules/notebooks/db/notebook_repository.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import Query
 
 from dataall.db import paginate
 from dataall.modules.notebooks.db.models import SagemakerNotebook
-from dataall.core.group.services.group_resource_manager import EnvironmentResource
+from dataall.core.group.services.environment_resource_manager import EnvironmentResource
 
 
 class NotebookRepository(EnvironmentResource):
